+++ PPM.hpp
@@ -12,6 +12,6 @@ class PPM
         PPM(int image_size);
         PPM(int image_size, unsigned char * image);
         void createImage(const char * name);
-        void setPixel(int x, int y, float color);
+        void setPixel(int x, int y, int color);
         float getPixel(int x, int y);
 };
